Se corrige fecha

+++ system/helpers/date_helper.php
@@ -389,12 +389,12 @@ if ( ! function_exists('mysql_to_unix'))
 
 		// YYYYMMDDHHMMSS
 		return mktime(
-			substr($time, 8, 2),
-			substr($time, 10, 2),
-			substr($time, 12, 2),
-			substr($time, 4, 2),
-			substr($time, 6, 2),
-			substr($time, 0, 4)
+			(int) substr($time, 8, 2),
+			(int) substr($time, 10, 2),
+			(int) substr($time, 12, 2),
+			(int) substr($time, 4, 2),
+			(int) substr($time, 6, 2),
+			(int) substr($time, 0, 4)
 		);
 	}
 }
